server/requirements.txt
google-cloud-aiplatform
Flask
python-dotenv
google-generativeai
flask-cors
bs4
